Lagt inn behandling av trekk-responsmeldinger fra OppdragZ
